requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anthropic==0.32.0
    # via -r requirements.in
anyio==4.4.0
    # via
    #   anthropic
    #   httpx
    #   openai
certifi==2024.7.4
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
distro==1.9.0
    # via
    #   anthropic
    #   openai
filelock==3.15.4
    # via huggingface-hub
fsspec==2024.6.1
    # via huggingface-hub
grpcio==1.65.2
    # via
    #   -r requirements.in
    #   grpcio-tools
grpcio-tools==1.65.2
    # via -r requirements.in
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httpx==0.27.0
    # via
    #   -r requirements.in
    #   anthropic
    #   openai
huggingface-hub==0.24.5
    # via tokenizers
idna==3.7
    # via
    #   anyio
    #   httpx
    #   requests
jiter==0.5.0
    # via anthropic
openai==1.37.1
    # via -r requirements.in
packaging==24.1
    # via huggingface-hub
protobuf==5.27.2
    # via
    #   -r requirements.in
    #   grpcio-tools
pydantic==2.8.2
    # via
    #   anthropic
    #   openai
pydantic-core==2.20.1
    # via pydantic
pymupdf==1.24.9
    # via -r requirements.in
pymupdfb==1.24.9
    # via pymupdf
pyyaml==6.0.1
    # via huggingface-hub
requests==2.32.3
    # via huggingface-hub
setuptools==72.1.0
    # via grpcio-tools
sniffio==1.3.1
    # via
    #   anthropic
    #   anyio
    #   httpx
    #   openai
tokenizers==0.19.1
    # via anthropic
tqdm==4.66.4
    # via
    #   huggingface-hub
    #   openai
typing-extensions==4.12.2
    # via
    #   anthropic
    #   huggingface-hub
    #   openai
    #   pydantic
    #   pydantic-core
urllib3==2.2.2
    # via requests
watchdog==4.0.1
    # via -r requirements.in
